requirements.txt
Flask==2.3.3
requests==2.32.3
Pillow==10.4.0
